Moved useInputState to utils folder
